requirements.txt
streamlit
matplotlib
numpy
requests
python-dateutil
httpx
certifi

# OpenAI + Pinecone
openai
pinecone

# Your project code dependencies
protobuf
typing_extensions

